+++ ci/module.f.ts
@@ -1,5 +1,5 @@
 import { utf8 } from '../text/module.f.ts'
-import { type NodeOperations, writeFile, type NodeEffect } from '../types/effect/node/module.f.ts'
+import { writeFile, type NodeEffect } from '../types/effect/node/module.f.ts'
 
 const os = ['ubuntu', 'macos', 'windows'] as const