+++ types/asn.1/test.f.ts
@@ -1,22 +1,45 @@
-import { empty, length, msb, unpack, vec, vec8, type Vec } from "../bit_vec/module.f.ts"
+import { empty, length, listToVec, msb, unpack, vec, vec8, type Vec } from "../bit_vec/module.f.ts"
 import { asBase } from "../nominal/module.f.ts"
-import { decode, encode, integer } from "./module.f.ts"
+import { decodeRaw, decodeInteger, encodeRaw, encodeInteger, integer, type Record, encode, decode, constructedSequence, octetString, boolean, constructedSet } from "./module.f.ts"
 
 const { concat, popFront: pop } = msb
+const cat = listToVec(msb)
 const pop8 = pop(8n)
 
 const check = (tag: number, v: Vec, rest: Vec) => {
-        const s = encode(tag)(v)
-        const [t0, v0, r] = decode(concat(s)(rest))
+        const s = encodeRaw([tag, v])
+        const [[t0, v0], r] = decodeRaw(concat(s)(rest))
         if (t0 !== tag) { throw `t0: ${t0}` }
         if (v0 !== v) { throw `v0: ${asBase(v0)}` }
         if (r !== rest) { throw `r: ${asBase(r)}` }
 }
 
+const integerValueCheck = (i: bigint, v: Vec) => {
+    const v0 = encodeInteger(i)
+    if (v !== v0) { throw `encode: ${asBase(v)}, ${asBase(v0)}` }
+    const i0 = decodeInteger(v)
+    if (i !== i0) { throw [i, i0] }
+}
+
+const ch0 = (r: Record, v: Vec, rest: Vec) => {
+    const [r0, rest0] = decode(concat(v)(rest))
+    if (rest0 !== rest) { throw `rest: ${asBase(rest0)}` }
+    const v0 = encode(r)
+    const v1 = encode(r0)
+    if (v !== v0) { throw `encode: ${asBase(v)}, ${asBase(v0)}` }
+    if (v !== v1) { throw `encode: ${asBase(v)}, ${asBase(v1)}` }
+}
+
+const ch = (r: Record, v: Vec) => {
+    ch0(r, v, empty)
+    ch0(r, v, vec8(0x23n))
+    ch0(r, v, vec(16n)(0x2345n))
+}
+
 export default {
     encodeSmall: () => {
         const v = vec8(0x13n)
-        const x = encode(integer)(v)
+        const x = encodeRaw([integer, v])
         const lx = length(x)
         if (lx !== 24n) { throw lx }
         const [tag, x1] = pop8(x)
@@ -31,7 +54,7 @@ export default {
         const valueLen = 0x7Fn << 3n
         const value = 0x1234n
         const v = vec(valueLen)(value)
-        const x = encode(integer)(v)
+        const x = encodeRaw([integer, v])
         const lx = length(x)
         if (lx !== 0x81n << 3n) { throw lx }
         const [tag, x1] = pop8(x)
@@ -46,7 +69,7 @@ export default {
         const valueLen = 0x80n << 3n
         const value = 0x123456n
         const v = vec(valueLen)(value)
-        const x = encode(integer)(v)
+        const x = encodeRaw([integer, v])
         const lx = length(x)
         if (lx !== 0x83n << 3n) { throw lx }
         const [tag, x1] = pop8(x)
@@ -63,7 +86,7 @@ export default {
         const valueLen = 0xFFn << 3n
         const value = 0x123456n
         const v = vec(valueLen)(value)
-        const x = encode(integer)(v)
+        const x = encodeRaw([integer, v])
         const lx = length(x)
         if (lx !== valueLen + (3n << 3n)) { throw `lx: ${lx}` }
         const [tag, x1] = pop8(x)
@@ -80,7 +103,7 @@ export default {
         const valueLen = 0x103n << 3n
         const value = 0x123456n
         const v = vec(valueLen)(value)
-        const x = encode(integer)(v)
+        const x = encodeRaw([integer, v])
         const lx = length(x)
         if (lx !== valueLen + (4n << 3n)) { throw `lx: ${lx}` }
         const [tag, x1] = pop8(x)
@@ -136,5 +159,50 @@ export default {
             //x40000000: () => check(integer, vec(0x4000_0000n)(0x8234n), vec8(0x23n)),
             // check(integer, vec(0x1_0000_0000n)(0x8234n), vec8(0x23n))
         },
-    ]
+    ],
+    integerValue: {
+        zero: () => integerValueCheck(0n, vec8(0n)),
+        one: () => integerValueCheck(1n, vec8(1n)),
+        minusOne: () => integerValueCheck(-1n, vec8(0xFFn)),
+        x7F: () => integerValueCheck(0x7Fn, vec8(0x7Fn)),
+        x80: () => integerValueCheck(0x80n, vec(16n)(0x80n)),
+        xFF: () => integerValueCheck(0xFFn, vec(16n)(0xFFn)),
+        nx7F: () => integerValueCheck(-0x7Fn, vec8(0x81n)),
+        nx80: () => integerValueCheck(-0x80n, vec8(0x80n)),
+        nx81: () => integerValueCheck(-0x81n, vec(16n)(0xFF7Fn)),
+        nx7FFF: () => integerValueCheck(-0x7FFFn, vec(16n)(0x8001n)),
+        nx8000: () => integerValueCheck(-0x8000n, vec(16n)(0x8000n)),
+        nx8001: () => integerValueCheck(-0x8001n, vec(24n)(0xFF7FFFn)),
+    },
+    encodeDecode: {
+        integer: () => {
+            ch([integer, 0n], cat([vec8(BigInt(integer)), vec8(1n), vec8(0n)]))
+            ch([integer, 1n], cat([vec8(BigInt(integer)), vec8(1n), vec8(1n)]))
+        },
+        sequence: () => {
+            ch([constructedSequence, []], cat([vec8(BigInt(constructedSequence)), vec8(0n)]))
+            ch(
+                [constructedSequence, [[integer, 0n]]],
+                cat([vec8(BigInt(constructedSequence)), vec8(3n), encode([integer, 0n])])
+            )
+            ch(
+                [constructedSequence, [[integer, 1n], [integer, 2n]]],
+                cat([vec8(BigInt(constructedSequence)), vec8(6n), encode([integer, 1n]), encode([integer, 2n])])
+            )
+            ch(
+                [constructedSequence, [[octetString, vec8(0x23n)], [boolean, true], [boolean, false]]],
+                cat([
+                    vec8(BigInt(constructedSequence)),
+                    vec8(9n),
+                    encode([octetString, vec8(0x23n)]),
+                    encode([boolean, true]),
+                    encode([boolean, false]),
+                ])
+            )
+        },
+        set: () => {
+            ch([constructedSet, [[integer, 2n], [integer, 1n]]],
+                cat([vec8(BigInt(constructedSet)), vec8(6n), encode([integer, 1n]), encode([integer, 2n])]))
+        }
+    }
 }